% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseDatasets.R, R/listDatasets.R
\name{browseDatasets}
\alias{browseDatasets}
\alias{listDatasets}
\title{List datasets in an HTML Browser}
\usage{
browseDatasets(drop.defaults = FALSE)

listDatasets(drop.defaults = FALSE)
}
\description{
List datasets in an HTML Browser

List datasets in a spreadsheet
}

