% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets}
\alias{datasets}
\title{View datasets from all installed package}
\usage{
datasets(package = NULL, lib.loc = NULL, all = TRUE,
  drop.defaults = FALSE)
}
\description{
View datasets from all installed package
}

