% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatasetInfo.R
\name{getDatasetInfo}
\alias{getDatasetInfo}
\title{Get information on installed datasets.}
\usage{
getDatasetInfo(package = NULL, lib.loc = NULL, all = TRUE)
}
\description{
Get information on installed datasets.
}

