% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseDatasets.R
\name{print.browseDatasets}
\alias{print.browseDatasets}
\title{Print function for browseDatasets objects}
\usage{
\method{print}{browseDatasets}(x, ...)
}
\arguments{
\item{x}{Object of class datasets}

\item{...}{Further arguments, ignored by the print method.}
}
\description{
Print function for browseDatasets objects
}
