% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.datasets.R
\name{print.datasets}
\alias{print.datasets}
\title{Print function for datasets class objects}
\usage{
\method{print}{datasets}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print function for datasets class objects
}
