% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printDatasets.R
\name{printDatasets}
\alias{printDatasets}
\title{Print datasets in the console}
\usage{
printDatasets(package = NULL, lib.loc = NULL, all = TRUE,
  drop.defaults = FALSE)
}
\description{
Print datasets in the console
}

