% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promtDatasets.R
\name{promptDatasets}
\alias{promptDatasets}
\title{Prompt for dataset in the console}
\usage{
promptDatasets(package = NULL, lib.loc = NULL, all = TRUE,
  drop.defaults = FALSE)
}
\description{
Prompt for dataset in the console
}

