% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDatasets.R
\name{viewDatasets}
\alias{viewDatasets}
\title{View datasets in a spreadsheet}
\usage{
viewDatasets(package = NULL, lib.loc = NULL, all = TRUE,
  drop.defaults = FALSE)
}
\description{
View datasets in a spreadsheet
}

