% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseDatasets.R
\name{browseDatasets}
\alias{browseDatasets}
\title{List datasets in an HTML Browser}
\usage{
browseDatasets(package = NULL, lib.loc = NULL, all = TRUE,
  drop.defaults = FALSE)
}
\description{
List datasets in an HTML Browser
}

