% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printDatasets.R
\name{printDatasets}
\alias{printDatasets}
\title{Print datasets in the console}
\usage{
printDatasets(
  package = NULL,
  lib.loc = NULL,
  all = TRUE,
  drop.defaults = FALSE
)
}
\arguments{
\item{package}{a character vector with the names of packages to search through, or NULL in which "all" packages (as defined by argument all) are searched.}

\item{lib.loc}{a character vector of directory names of R libraries, or NULL. The default value of NULL corresponds to all libraries currently known.}

\item{all}{logical; if TRUE search all available packages in the library trees specified by lib.loc, and if FALSE, search only attached packages.}

\item{drop.defaults}{logical; if TRUE, do not include the datasets from the datasets package.}
}
\description{
Print datasets in the console
}
